requirements.txt
Django==4.1.3
djangorestframework==3.14.0
requests==2.31.0
pygsheets==2.0.6
line-bot-sdk==3.7.0
django-environ==0.9.0